Optimized collision in game
The collision between the the player and all of the objects is something
that still needs some refactoring for sure.  its a bit messy and should
be rewritten from the ground up.  That said I did remove several uneeded
lines of code to make the existing version a little more efficient and a
lot cleaner.